Last_ver
Last_ver

+++ bot_diplom.py
@@ -27,29 +27,38 @@ logger = logging.getLogger(__name__)
 
 FIRST, SECOND, THIRD, FOURTH = range(4)
 
+conn = mysql.connector.connect(host='mysql.j949396.myjino.ru', 
+                               database='j949396', 
+                               user='j949396', 
+                               password='qwerty')
+cursor = conn.cursor()
+
 
 def talk_to_me(bot, update):
     # функция - ответ на текст введенный пользователем
     update.message.reply_text('Нажмите /start для запуска бота')
 
-def greet_user(bot, update):
+def greet_user(bot, update, user_data):
     # функция - /start
-    text = 'Вас приветствует salon_service_bot!'
-    my_keyboard = ReplyKeyboardMarkup([['Запись'],
-                                       ['Мои записи', 'О нас']],
-                                      resize_keyboard=True,
-                                      one_time_keyboard=True)
-    update.message.reply_text(text, reply_markup=my_keyboard)
+    if user_data == {}:
+        text = 'Вас приветствует salon_service_bot!'
+        my_keyboard = ReplyKeyboardMarkup([['Запись'],
+                                           ['Мои записи', 'О нас']],
+                                          resize_keyboard=True,
+                                          one_time_keyboard=True)
+        update.message.reply_text(text, reply_markup=my_keyboard)
+    else: 
+        text_2 = 'Выберите дальнейшее действие'
+        my_keyboard_2 = ReplyKeyboardMarkup([['Мои записи'],
+                                   ['Отменить запись', 'Добавить Запись']],
+                                  resize_keyboard=True,
+                                  one_time_keyboard=True)
+        update.message.reply_text(text_2, reply_markup=my_keyboard_2)
 
 
-def choose_master(bot, update):
-    # функция вызова инлайн клавиатуры с мастерами
-    conn = mysql.connector.connect(host='mysql.j949396.myjino.ru', 
-                                   database='j949396', 
-                                   user='j949396', 
-                                   password='qwerty')
-    cursor = conn.cursor()
-    sql = "SELECT barber_name FROM barbers"
+def choose_master(bot, update, user_data):
+    # функция вызова инлайн клавиатуры с мастерами 
+    sql = "SELECT service_name FROM services"
     cursor.execute(sql)
     data_base = cursor.fetchall()
     all_masters = []
@@ -57,22 +66,22 @@ def choose_master(bot, update):
         all_masters.append(masters[0])
     keyboard = []
     row = []
-    for i in all_masters:
+    rows = []
+    for i in all_masters[0:2]:
         row.append(InlineKeyboardButton(i, callback_data=str(i)))
+    for i in all_masters[2:4]:
+        rows.append(InlineKeyboardButton(i, callback_data=str(i)))
+            
     keyboard.append(row)
+    keyboard.append(rows)
     reply_markup = InlineKeyboardMarkup(keyboard)
     # bot.send_photo(chat_id=update.message.chat.id,
     #                photo=open('C:\projects\diplom\photo\BRB 666.jpg', 'rb'))
-    update.message.reply_text('Выберите мастера', reply_markup=reply_markup)
+    update.message.reply_text('Выберите услугу', reply_markup=reply_markup)
     return FIRST
 
 def choose_service(bot,update, user_data):
     # функция вызова инлайн клавиатуры с услугами
-    conn = conn = mysql.connector.connect(host='mysql.j949396.myjino.ru', 
-                                          database='j949396', 
-                                          user='j949396', 
-                                          password='qwerty')
-    cursor = conn.cursor()
     sql = "SELECT * FROM barbers"
     cursor.execute(sql)
     data_base = cursor.fetchall()
@@ -87,35 +96,33 @@ def choose_service(bot,update, user_data):
 
     query = update.callback_query
     name = query.data
-# Клавиатура с услугами
+    # Клавиатура с услугами
     counter = []
-    for masters in data_base:
-        if name in masters:
-            global c
-            c = query.data
-            a = masters[0]
-            for master_id in data_base_1:
-                if a in master_id:
-                    b = master_id[2]
-                    for service_id in data_base_2:
-                        if b in service_id:
-                            all_services = []
-                            all_services.append(service_id[2])
+    for service_id in data_base_2:
+        if name in service_id:
+            a = service_id[0] 
+            for barber_id in data_base_1:
+                if a in barber_id:
+                    b = barber_id[1]
+                    for barber_name in data_base:
+                        if b in barber_name:
+                            all_barbers = []
+                            all_barbers.append(barber_name[1])
                             keyboard = []
                             row = []
-                            for i in all_services:
+                            for i in all_barbers:
                                 row.append(InlineKeyboardButton(i, callback_data=str(i)))
                             counter = row + counter
     list_1 = []
     list_1.append(counter)
     reply_markup = InlineKeyboardMarkup(list_1)
 
-    bot.edit_message_text(text='Выберите услугу:',
+    bot.edit_message_text(text='Выберите мастера:',
                           chat_id=update.callback_query.from_user.id,
                           message_id=query.message.message_id,
                           reply_markup=reply_markup)
 # Запись данных в user_data
-    user_data ['name'] = query.data
+    user_data ['service'] = query.data
     return SECOND
 
 def calendar(bot,update, user_data):
@@ -127,43 +134,73 @@ def calendar(bot,update, user_data):
     bot.edit_message_reply_markup(chat_id=query.message.chat_id,
                                   message_id=query.message.message_id,
                                   reply_markup=telegramcalendar.create_calendar())
-    user_data ['service'] = query.data
+    user_data ['name'] = query.data
     return THIRD
 
 def time(bot,update, user_data):
     # функция вызова инлайн клавиатуры с временем
     query = update.callback_query
-    selected,date = telegramcalendar.process_calendar_selection(bot, update)
+    selected, date = telegramcalendar.process_calendar_selection(bot, update)
+    
+    sql = "SELECT * FROM time_to_barbers"
+    cursor.execute(sql)
+    data_base = cursor.fetchall()
+
+    sql_2 = "SELECT * FROM record_info"
+    cursor.execute(sql_2)
+    data_base_2 = cursor.fetchall()
+
+    # проверка по записанным клиентам
     if selected:
-        inline_keyboard = [[InlineKeyboardButton('10:00', callback_data ='10:00'),
-                            InlineKeyboardButton('11:00', callback_data ='11:00')],
-                            [InlineKeyboardButton('12:00', callback_data ='12:00'),
-                            InlineKeyboardButton('13:00', callback_data ='13:00')],
-                            [InlineKeyboardButton('14:00', callback_data ='14:00'),
-                            InlineKeyboardButton('15:00', callback_data ='15:00')]]
-        reply_markup = InlineKeyboardMarkup(inline_keyboard)
+        all_time = []
+        for name_list in data_base:
+            if user_data.get('name') in name_list[0]:
+                all_time.append(name_list[1])
+        for time_info in data_base_2:
+            if user_data.get('name') in time_info[1]:
+                if date.strftime("%Y-%m-%d") in time_info[2]:
+                    if time_info[3] in all_time:
+                        all_time.remove(time_info[3])
+                        if len(all_time) < 6:
+                            all_time.append('Нет записи')
+
+        keyboard = []
+        row = []
+        row_2 = []
+        for i in all_time[0:3]:
+            row.append(InlineKeyboardButton(i, callback_data=str(i)))
+        for i in all_time[3:6]:
+            row_2.append(InlineKeyboardButton(i, callback_data=str(i)))
+
+        keyboard.append(row)
+        keyboard.append(row_2)
+        reply_markup = InlineKeyboardMarkup(keyboard)
         bot.edit_message_text(text='Выберите время:',
-                          chat_id=update.callback_query.from_user.id,
-                          message_id=query.message.message_id)
+                              chat_id=update.callback_query.from_user.id,
+                              message_id=query.message.message_id)
         bot.edit_message_reply_markup(chat_id=query.message.chat_id,
                                       message_id=query.message.message_id,
                                       reply_markup=reply_markup)
     user_data ['date'] = date.strftime("%Y-%m-%d")
     return FOURTH
 
+
 def contact (bot,update, user_data):
     # функция вызова запроса контактов
     query = update.callback_query
-    contact_button = KeyboardButton('Контактные данные', request_contact=True)
-    my_keyboard = ReplyKeyboardMarkup([[contact_button]],
-                                      resize_keyboard=True,
-                                      one_time_keyboard=True)
-    bot.send_message(chat_id=update.callback_query.from_user.id,
-                     text="Отправьте Ваши контактные данные для уточнения заказа:",
-                     reply_markup=my_keyboard)
-    bot.delete_message(chat_id=update.callback_query.from_user.id,
-                       message_id=query.message.message_id)
-    user_data ['time'] = query.data
+    if query.data == "Нет записи":
+        bot.answer_callback_query(callback_query_id= query.id)
+    else:
+        contact_button = KeyboardButton('Контактные данные', request_contact=True)
+        my_keyboard = ReplyKeyboardMarkup([[contact_button]],
+                                          resize_keyboard=True,
+                                          one_time_keyboard=True)
+        bot.send_message(chat_id=update.callback_query.from_user.id,
+                         text="Отправьте Ваши контактные данные для уточнения заказа:",
+                         reply_markup=my_keyboard)
+        bot.delete_message(chat_id=update.callback_query.from_user.id,
+                           message_id=query.message.message_id)
+        user_data ['time'] = query.data
 
 
 def get_contact(bot, update, user_data):
@@ -194,21 +231,15 @@ def get_contact(bot, update, user_data):
 
 
 # Запись всех данных в БД
-    # conn = mysql.connector.connect(host='mysql.j949396.myjino.ru', 
-    #                                database='j949396', 
-    #                                user='j949396', 
-    #                                password='qwerty')
-    # cursor = conn.cursor()
-    # cort_1 = (user_data.get('name'),)
-    # cort_2 = cort_1 + (user_data.get('service'),)
-    # cort_3 = cort_2 + (user_data.get('date'),)
-    # cort_4 = cort_3 + (user_data.get('time'),)
-    # cort_5 = cort_4 + (user_data.get('phone'),)
-    # print(cort_5)
-    # data = []
-    # data.append(cort_5)
-    # cursor.execute("INSERT INTO record_info (name, service, date, time, number) VALUES (%s, %s, %s, %s, %s)", cort_5)
-    # conn.commit()
+    cort_1 = (user_data.get('name'),)
+    cort_2 = cort_1 + (user_data.get('service'),)
+    cort_3 = cort_2 + (user_data.get('date'),)
+    cort_4 = cort_3 + (user_data.get('time'),)
+    cort_5 = cort_4 + (user_data.get('phone'),)
+    data = []
+    data.append(cort_5)
+    cursor.execute("INSERT INTO record_info (name, service, date, time, number) VALUES (%s, %s, %s, %s, %s)", cort_5)
+    conn.commit()
     # cursor.close()
     # conn.close()
     
@@ -220,7 +251,8 @@ def my_entry(bot, update, user_data):
     update.message.reply_text("Имя мастера: " + user_data.get('name') + "\n"
                               "Услуга: " + user_data.get('service') + "\n"
                               "Дата: " + user_data.get('date') + "\n"
-                              "Время: " + user_data.get('time'))
+                              "Время: " + user_data.get('time'),
+                              reply_markup=my_keyboard_2)
 
 
 def info(bot, update):
@@ -238,31 +270,51 @@ def info(bot, update):
                               reply_markup=my_keyboard)
 
 
+def delete(bot, update, user_data):
+    sql = "SELECT number FROM record_info"
+    cursor.execute(sql)
+    data_base = cursor.fetchall()
+    for phone in data_base:
+        a = user_data.get('phone')
+        cursor.execute("DELETE FROM record_info WHERE number =  %s" % a)
+        conn.commit()
+    user_data.clear()
+    greet_user(bot, update, user_data)
+
+
 def main():
     mybot = Updater("728852231:AAEZLnITK0BYNpAfQ4DCIC8CjpyiYLYUpIo", request_kwargs=PROXY)
     
     dp = mybot.dispatcher
 
     conv_handler = ConversationHandler(
-        entry_points=[RegexHandler('Запись', choose_master)],
+        entry_points=[RegexHandler('Запись', choose_master, pass_user_data=True),
+                      RegexHandler('Добавить Запись', choose_master, pass_user_data=True)],
         states={FIRST: [CallbackQueryHandler(choose_service, pass_user_data=True)],
                 SECOND: [CallbackQueryHandler(calendar, pass_user_data=True)],
                 THIRD: [CallbackQueryHandler(time, pass_user_data=True)],
                 FOURTH: [CallbackQueryHandler(contact, pass_user_data=True)]},
-        fallbacks=[MessageHandler(Filters.contact, get_contact, pass_user_data=True)])
-    
-    dp.add_handler(CommandHandler("start", greet_user))
-    
+        fallbacks=[MessageHandler(Filters.contact, get_contact, pass_user_data=True)],
+        allow_reentry=True)
+
     dp.add_handler(conv_handler)
+    
+    dp.add_handler(CommandHandler("start", greet_user, pass_user_data=True))
     
     dp.add_handler(CommandHandler("Мои записи", my_entry, pass_user_data=True))
     dp.add_handler(RegexHandler("Мои записи", my_entry, pass_user_data=True))
 
+    dp.add_handler(CommandHandler("Отменить запись", delete, pass_user_data=True))
+    dp.add_handler(RegexHandler("Отменить запись", delete, pass_user_data=True))
+
+    dp.add_handler(CommandHandler("Вернуться в главное меню", greet_user, pass_user_data=True))
+    dp.add_handler(RegexHandler("Вернуться в главное меню", greet_user, pass_user_data=True))
+
+    dp.add_handler(CommandHandler('Запись', choose_master, pass_user_data=True))
+    dp.add_handler(RegexHandler('Запись', choose_master, pass_user_data=True))
+
     dp.add_handler(CommandHandler("О нас", info))
     dp.add_handler(RegexHandler("О нас", info))
-
-    dp.add_handler(CommandHandler("Вернуться в главное меню", greet_user))
-    dp.add_handler(RegexHandler("Вернуться в главное меню", greet_user))
 
     dp.add_handler(MessageHandler(Filters.text, talk_to_me))
